requirements.txt
crispy-bootstrap5==2025.4
Django==5.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.3.0
django-crispy-forms==2.4
djangorestframework==3.14.0